+++ src/lib/kotlin/slatekit-apis/src/main/kotlin/slatekit/apis/tools/code/Templates.kt
@@ -4,7 +4,10 @@ import slatekit.common.io.Uris
 import slatekit.common.pascalCase
 import java.io.File
 
-class CodeGenTemplates(val templates: List<CodeGenTemplate>) {
+/**
+Wrapper for the 3 code generation templates for creating the APIs and DTOs
+ */
+class Templates(val templates: List<CodeGenTemplate>) {
 
     val api = templates.first { it.name.startsWith("api.") }
 
@@ -13,13 +16,13 @@ class CodeGenTemplates(val templates: List<CodeGenTemplate>) {
     val dto = templates.first { it.name.startsWith("model.") }
 
     companion object {
-        fun load(path: String, lang: Language): CodeGenTemplates {
+        fun load(path: String, lang: Language): Templates {
             val templates = listOf(
                  getContent(path, "api.${lang.ext}"),
                  getContent(path, "method.${lang.ext}"),
                  getContent(path, "model.${lang.ext}")
             ).map { CodeGenTemplate(it.first, it.second, it.third) }
-            return CodeGenTemplates(templates)
+            return Templates(templates)
         }
 
 
@@ -34,6 +37,9 @@ class CodeGenTemplates(val templates: List<CodeGenTemplate>) {
 
 class CodeGenTemplate(val path:String, val name:String, val raw: String) {
 
+    /**
+     * Generates a file using the template, variables provided and file name
+     */
     fun generate(vars: Map<String, String>, targetFolder:String, targetName:String, lang: Language) {
         val replacements = vars.map { "@{${it.key}}" to it.value }
         val template = replacements.fold(raw){ acc, next ->
